backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
base58==2.1.1
bcrypt==4.3.0
bitarray==3.3.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cytoolz==1.0.1
ecdsa==0.19.1
eth-account==0.10.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.2.2
fastapi==0.109.0
frozenlist==1.5.0
h11==0.16.0
hexbytes==0.3.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
ipfshttpclient==0.8.0a2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lru-dict==1.2.0
multiaddr==0.0.9
multidict==6.4.3
netaddr==1.3.0
packaging==25.0
parsimonious==0.10.0
passlib==1.7.4
pluggy==1.5.0
propcache==0.3.1
protobuf==6.30.2
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
pytest==7.4.3
pytest-asyncio==0.23.5
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pyunormalize==16.0.0
# pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
tomli==2.2.1
toolz==1.0.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.0
varint==1.0.2
watchfiles==1.0.5
web3==6.12.0
websockets==15.0.1
yarl==1.19.0
